app/requirements.txt
GPUtil
scikit-learn
flask
celery
requests
pyelftools
loguru
psutil
tensorflow
joblib
numpy
cryptography
torch torchvision torchaudio
torch
pynvml
azure-identity
azure-mgmt-resource
azure-mgmt-monitor
azure-sentinel
azure-loganalytics
azure-mgmt-security
azure-mgmt-network
azure-mgmt-trafficanalytics
azure-resourcegraph

